re-organizing addition of github commit build arg parser
currently the main function is a bit hairy, so this change moves the functions into
their respective parsers (and adds a few notes in the docs for how to use it)
