+++ tests/ubuntu/21.04/Dockerfile
@@ -1,14 +1,14 @@
 LABEL maintainer vsoch
 
 # Testing a github release update
-ARG uptodate_github_release_spack__spack=v0.16.2
+ARG uptodate_github_release_spack__spack=v0.16.1
 ARG uptodate_github_release_spack__spack=
 
 # Testing github commit
-ARG uptodate_github_commit_spack__spack__develop=5d23638fdc3d93c70b26b5f8ab37ba51e8f707e6
+ARG uptodate_github_commit_spack__spack__develop=NA
 ARG uptodate_github_commit_spack__spack__develop=
 
 # Testing a spack version update
-ARG uptodate_spack_ace=6.5.12
+ARG uptodate_spack_ace=6.5.6
 
 RUN apt-get update && apt-get install -y vim
